Simplify element sanitizing
This method simply loops through every single element and deletes it where appropriate. For now an XPath query is still used to iterating through elements straightforward; this may change later.

This should simplify handling XHTML (where arbitrary namespaces can be involved), and will ensure that every single element in the document is handled consistently.

There do, however, remain some cases that are not handled intelligently, such as if the root element is not in the keep list.